extracts spam protected duplicated logic into a form class
