+++ lib/developer_portal/app/views/developer_portal/password/new.html.liquid
@@ -6,7 +6,7 @@
         Forgot password
       </div>
       <div class="panel-body">
-        {% form 'password_reset', account, class: 'form-horizontal' %}
+        {% form 'password_reset', { class: 'form-horizontal' } %}
           <fieldset>
             <div class="form-group">
               <label for="email" class="control-label col-md-4">Email address</label>
